WT-7965 Use the same write generation for a page that is written during recovery
During WiredTiger recovery, all the existing transaction ids are reset
based on the page write generation numbers present on the page after
rollback to stable is sucessfully finished. To avoid having transction ids
on the pages that are written during recovery, WiredTiger uses the same
write generation number of the page that exists for the new page image
that is going write to disk. With this approach, when this page gets
loaded during the process of rollback to stable, it preseves the transction
ids and resets only upon next restart.