Slice out the last line in test_verbose01
Given that the JSON messages have significantly more characters,
we cant always rely that the maximum number of characters will be
enough to avoid reading a truncated message. To make this test
more resilient, we should slice out the last message string in the
off chance that it is truncated ('...').