Force the test_util+windows_shim libs to be static
Given these libraries are strongly tied to our C/C++ tests, theres
no need to compile these libraries as shared targets i.e. they aren't
going to be used beyond the context of this project. We can afford
to compile these libraries as static, simplifying the dependencies
of the various test executables that link against it.