Pass 'EXT_LIBPATH' to test/format
Prior we were compiling the full path of each ext module into the
test format binary. This approach breaks when executing test
format on a different host. To fix this, fallback to using
relative paths. To accomodate CMake builds in this case, we
expose the CPP variable 'EXT_LIBPATH', as means to add or drop
the '.libs' subdir.